Move asset classes from root to a separate package
