+++ flixelgdx-html5/src/main/java/org/flixelgdx/backend/html5/asset/FlixelHtml5AssetManager.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.flixelgdx.backend.html5;
+package org.flixelgdx.backend.html5.asset;
 
 import org.flixelgdx.Flixel;
 import org.flixelgdx.asset.FlixelAsset;
